+++ src/Shipment/resources/database/migrations/2023_02_20_094957_add_reference_number_to_shipments_table.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+return new class () extends Migration {
+    public function up()
+    {
+        Schema::table('shipments', function (Blueprint $table) {
+            $table->string('reference_number')->nullable()->after('tracking_number');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('shipments', function (Blueprint $table) {
+            $table->dropColumn('reference_number');
+        });
+    }
+};