+++ Changelog.md
@@ -7,6 +7,7 @@
 
 - Added the `description` and `state` fields to the master product variants table
 - Added the Address comparison utility
+- Added the `reference_number` field to the shipments model
 - Fixed the `Carrier::name()` method to overcome accessor infinite loop
 - Fixed the address creation with order factory when passing an explicit address type